+++ Output Files/WinOutputs.h
@@ -74,7 +74,7 @@ protected:
 	 *
 	 * Sends the appropriate output message to all registered clients.
 	 */
-	void SendOutput(EOutputs output, UINT8 prevValue, UINT8 value);
+	void SendOutput(EOutputs output, UINT32 prevValue, UINT32 value);
 
 private:
 	static bool s_createdClass;
